правит тест с избранными
