+++ src/store/slices/favorites.ts
@@ -33,8 +33,17 @@ const favoritesSlice = createSlice({
     builder.addCase(toggleFavoriteAction.pending, (state) => {
       state.statusToggleFavorite = RequestStatus.Loading;
     });
-    builder.addCase(toggleFavoriteAction.fulfilled, (state) => {
+    builder.addCase(toggleFavoriteAction.fulfilled, (state, action) => {
       state.statusToggleFavorite = RequestStatus.Succeed;
+      const changedOffer = action.payload;
+
+      if (changedOffer.isFavorite) {
+        state.favorites.push(changedOffer);
+        return;
+      }
+
+      const changedFavoriteIndex = state.favorites.findIndex((favoriteOffer) => favoriteOffer.id === changedOffer.id);
+      state.favorites.splice(changedFavoriteIndex, 1);
     });
     builder.addCase(toggleFavoriteAction.rejected, (state) => {
       state.statusToggleFavorite = RequestStatus.Failed;
